+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Improvements
 
+- [#10](https://github.com/MalteHerrmann/upgrade-helper/pull/10) Estimate the upgrade height based on last 50.000 blocks.
 - [#9](https://github.com/MalteHerrmann/upgrade-helper/pull/9) Add logic for selection and calculation of upgrade date and time.
 - [#5](https://github.com/MalteHerrmann/upgrade-helper/pull/5) Prepare proposal contents depending on input.
 - [#4](https://github.com/MalteHerrmann/upgrade-helper/pull/4) Check if release already exists on GitHub.